[PLAT-14900] Exclude empty string from regex validation for AZUoptional fields
Summary:
The optional resource group fields for AZU providers have regex pattern
validation.
Since this field is option, we should be ignoring the field when it is
empty.

Test Plan:
Verify with a name that fails the regex pattern validation (ex.
`resourceGroup.`). An error message is expected.
Verify with a name that passes the regex pattern validation (ex.
`resourceGroup`). No error message is expected.
Verify with no input in the optional resource group fields. No error message
is expected.

Reviewers: lsangappa, yshchetinin, rmadhavan

Reviewed By: lsangappa

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37249