[#22989] YSQL: Fix regular expression pushdown
Summary:
Regular expressions are compiled first and then executed. To avoid recompiling a regular expression each time, Postgres has implemented a cache (size 32) that stores the compiled regular expressions. This is implemented as a static array, and works well for PG because PG is single-threaded.

When expression pushdown is enabled in Yugabyte, Postgres expression evaluation is carried out on DocDB side. The same code is executed in a multi-threaded environment, so multiple threads are accessing the same objects. This can result in correctness issues (because the compiled regex is not what was expected) or tserver crashes (because the cache is corrupted).

Using a thread local cache solves the problem. Each compiled regex is malloc'd, so to handle its lifetime, a new object is created in C++ code that invokes the cleanup for the regular expressions in the cache.
Jira: DB-11908

Test Plan:
```
./yb_build.sh --cxx-test pgwrapper_pg_mini-test --gtest_filter PgMiniTest.RegexPushdown
```

fails on master, passes with this change.

Reviewers: amartsinchyk, dmitry

Reviewed By: dmitry

Subscribers: dmitry, smishra, yql

Differential Revision: https://phorge.dev.yugabyte.com/D36095