[#23478] YSQL: fix connection manager session variable case sensitivity issue
Summary:
The SET session variables for a client connection are persisted in the vars field in client and server context. These session variables in postgres are not case sensitive yet in the connection manager layer they are being persisted in a case sensitive manner.

The change here is to persist the session variables in connection manager in lower case and also the comparisons wherever needed are also done in lower case.

The implication of this change is that when the session variables with the same name but different case sensitivity are used, we will parse it and interpret it only in lower case.

For example, while we are running the pg regress tests, some environment variables are set such as `PGTZ`, `PGDATESTYLE` and `PGOPTIONS=-c intervalstyle=postgres_verbose`. These get translated to `timezone`, `datestyle` and options is parsed to `intervalstyle` correspondingly as part of the Startup packet parsing. This info is intercepted by connection manager and persisted in the client vars field. When these vars are sent to the backend postgres connection to set the client connection state, postgres returns them in a different case `timezone` as `TimeZone`, `datestyle` as `DateStyle` and `intervalstyle` as `IntervalStyle`.  Connection manager when it receives the ParameterStatus packet parses them and persists in the client vars and server vars entries. This is resulting in the duplication of the same postgres GUC variables in different cases. This can lead to different failure scenarios. One of the failure is mentioned below:

  - As part of the postgres backend initialization, the following GUC options are set in the manner
```
SET intervalstyle=E'postgres_verbose';SET application_name=E'pg_regress/yb_pg_init';SET datestyle=E'Postgres, MDY';SET timezone=E'PST8PDT';
```
  - Now, during subsequent SET operation `SET client_min_messages = warning;`, the ParameterStatus messages are persisted in client vars. The `datestyle` is returned as `DateStyle` by postgres and is persisted the same way. So, now in client vars two options with the same key are present with different case sensitivity and the server vars has `DateStyle` .
The client vars are

```
datestyle=Postgres, MDY
DateStyle = Postgres, MDY
```
server vars are

```
DateStyle = Postgres, MDY
```

  - Now, let us perform the next SET operation `SET DATESTYLE TO ISO;` . With this operation, the new client vars are
```
datestyle=Postgres, MDY
DateStyle = ISO, MDY
```
The server vars are
```
DateStyle = ISO, MDY
```
  - Now, if we try to perform any other operation using connection manager, the SET session variables command looks like as below
```
SET intervalstyle=E'postgres_verbose';SET application_name=E'pg_regress/yb_pg_orafce';SET datestyle=E'Postgres, MDY';SET timezone=E'PST8PDT';SET IntervalStyle=E'postgres_verbose';SET TimeZone=E'PST8PDT';SET client_min_messages=E'warning';
```
In the above set statements, we will notice that `datestyle=E'Postgres, MDY'` even though the prior SET command was `SET DATESTYLE TO ISO;` . This is because if the client and server var for a certain option are the same in a case sensitive manner, then that SET command is not applied before the actual query is run. Yet, `datestyle=E'Postgres, MDY'` is being SET  as the comparison is case sensitive.

  - To overcome this. I am persisting all the SET commands in lower case for client, server and also the comparisons wherever needed are done on lower case strings. So, all the SET option commands that are called before the Query will be called with lowercase option names.

**Changes made**
There are 3 places where the changes are made:
  # `kiwi_var_init` -> This method is called to initialize the kiwi_var struct. As part of this method, the name and name_len values are persisted. By the logic mentioned above, name would be converted to lowercase and persisted in `var->name.`
  # `yb_kiwi_var_push` -> This method is called to create a new entry in `kiwi_vars_t` struct. The new entry would be created with name in lowercase and the other fields are persisted as passed in the method. This method is primarily called by `kiwi_vars_update` which in turn is called while dealing with ParameterStatus or Startup packet messages. In both the cases as discussed above, we would like to persist the name in lowercase for correctness sake as explained in the example above.
  # `yb_kiwi_vars_get` -> This method is used to fetch the kiwi_var struct corresponding to the param name. As discussed above, the ParameterStatus response need not be in lower case and hence, the idea is to persist the entries in lower case and thereafter search them in lower case which will lead to consistent behaviour while dealing with session variables/GUC variables.

Jira: DB-12395

Test Plan:
Jenkins: test regex: .*TestPgRegress.*, enable connection manager

The tests `org.yb.pgsql.TestPgRegressMisc#testPgRegressMiscSerial2` and `org.yb.pgsql.TestPgRegressPushdown#testPgRegressPushdown` in pg regress test this feature quite extensively. The issue was noticed when these tests failed and the fix is to address them. So, no additional tests are needed.

Reviewers: asrinivasan, skumar, mkumar, rbarigidad, jason, stiwary

Reviewed By: skumar

Subscribers: svc_phabricator, jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D37295