Merged PR 1429564: adding support for passing window value in context
### Does this PR have any customer impact?

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)

### Are you introducing any new config? If yes, do you have tests with and without them being set?

### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description

adding support for passing extra window value
This window value is required in individual setWindowField operators like $rank, $denseRank, $locF,
As for each of them we need  a way to check if the windowValue is empty or specifically content
Hence, raising a common pr

----
#### AI description  (iteration 1)
#### PR Classification
New feature: adding support for passing window value in context.

#### PR Summary
This pull request introduces the ability to pass a window value within the context for window aggregation operations.
- `src/aggregation/bson_aggregation_window_operators.c`: Added `windowValue` to the context and updated the order of function calls to include `UpdateWindowAggregationOperator` with the new context.
- `src/aggregation/bson_aggregation_window_operators.c`: Introduced a new field `windowValue` in the `WindowOperatorContext` struct.