[#23393] YSQL, ASH: Fix TSAN test failures when ASH is enabled by default
Summary:
There are a lot of malloc and free calls from signal handlers, which is not
async-signal-safe. TSAN attempts to report this and gets stuck more
often than not when the start time of a postgres background worker is
BgWorkerStart_ConsistentState or BgWorkerStart_RecoveryFinished.

ASH uses BgWorkerStart_ConsistentState, and thus a lot of existing tests
fail in TSAN builds because postmaster gets stuck and the test cannot
complete in the given time.

This diff fixes the issue by using the TSAN flag report_signal_safe in
TSAN_OPTIONS, which was incorrectly used in ASAN_OPTIONS.
This causes TSAN to not report async-signal-unsafe code and thus
postmaster doesn't get stuck.

Note that this doesn't fix the original issue of postgres having
async-signal-unsafe code, this just makes sure that we don't see a
bunch of TSAN failures when ASH is enabled.

This diff also enables some of the existing ASH tests in TSAN mode.
Some of the tests from the wait state itest framework which use YCQL
are not enabled in TSAN because they are currently failing due to a
separate issue and will be fixed in a separate diff.
Jira: DB-12315

Test Plan:
Jenkins: build type: tsan

./yb_build.sh tsan --cxx-test wait_states-itest
./yb_build.sh tsan --cxx-test pg_mini-test --gtest-filter PgMiniTest.Ash

Reviewers: dmitry, jason, amitanand, hsunder

Reviewed By: dmitry, hsunder

Subscribers: hbhanawat, yql

Differential Revision: https://phorge.dev.yugabyte.com/D37049