[PLAT-14945][PLAT-15006][PLAT-15017]: Fix xCluster metrics and Add/Remove tables with bootstrap needed.
Summary:
xCluster metrics calculation fails for all if the universe of one of them is deleted and this results in failure of metrics for all xCluster config which happened to the dev portal.
To avoid the issue, we will check if xCluster is in running status before fetching universes.

Another issue this diff resolved is adding/removing the same table in a single API call where tableToAdd requires bootstrap.
There were these two following cases:
 - Add a table that requires bootstrap and remove a table that still exists:
    - Here, we were getting errors in the API itself as we set all database tables in bootstrap tables params if any table requires bootstrap. Later when during another ops, we were trying to get the tableInfoList of the table that is planned to be removed from replication but exists in bootstrap params from the filtered tableIdsInReplicationInfoList, the table info list of such tables will exist in sourceUniverseTableInfoList.
 - Add a table that requires Bootstrap and remove a table that is deleted from the source
    - We were getting issues while dropping the database as part of Bootstrap, as dropped tables from the source are not included in tableIdsDeleteReplication extracted from the current source universe table list.

Also, Earlier In YCQL if a table required bootstrap then we added all tables in the database for deletion but it is not necessary in the case of YCQL as we support backup & Restore on table level in YCQL, hence we will not be adding all the table in the database rather add parent and index tables if anyone requires bootstrap.

Test Plan:
Tested manually:
  - Verified that metrics are populated for other xCluster configs if an xCluster config is in a DeletedUniverse state.

  - Add a table that requires bootstrap and remove a table deleted on a 2 DB xCluster config.
  - Add a table that requires bootstrap and remove a table that is not deleted.

 - Add 2 tables where one requires bootstrap and the other does not in a ycql replication.
 - Add a table in ycql replication that requires bootstrap in a ycql replication
 - Add tables that do not require bootstrap in a ycql replication

Also verified below scenarios as requested:

 - Add 2 YSQL tables from two DBs where one requires bootstrap and the other does not and make sure the replication for one of the DBs is deleted and recreated through bootstrapping, while the other DB is intact and the new table is simply added to the replication without bootstrapping.
 - Add a YSQL table that requires bootstrap where the xCluster config has only one DB
 - Add a YSQL table that does not require bootstrap

Reviewers: hzare, cwang, sanketh

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37323