[PLAT-14863] Add validation for xCluster replication name
Summary:
This diff adds a validation check for illegal characters in xCluster
replication config name during the creation workflow.

Users will receive a validation error message which indicates which
characters are not allowed:
`The name of the replication configuration cannot contain any characters in [SPACE '_' '*' '<' '>' '?' '|' '\"' NULL])"`

Test Plan:
- Attempt to create an xCluster config with no replication name. Verify that
the UI marks the field as required.
- Attempt to create an xCLuster config using a replication name that
contains illegal characters (ex. space or _). Verify that the replication
name field is marked as errored and a message informs the user which
characters are no permitted.

{F274901}
{F274904}
{F274905}

Reviewers: rmadhavan, kkannan, cwang, hzare

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37217