[#23374] DocDB: Refactor ExecuteBatchYbctid to take keys from an abstract KeyProvider
Summary:
This change refactors the method on a main table scan that takes in ybctid's sourced from an index to abstract its source of point-lookup keys. This will allow us to add keys sourced from a secondary vector index as a key source.

Instead of `ExecuteBatchYbctid` taking in a vector of ybctid's, this method now takes in an abstract `KeyProvider` instance. There is only one implementation of this `KeyProvider` class `PgsqlReadRequestYbctidProvider` that provides ybctid's from the current request protobuf.
Jira: DB-12296

Test Plan: Jenkins

Reviewers: sergei

Reviewed By: sergei

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D37033