[PLAT-14951] Add positive interger error message to pitr param step
Summary:
**Context**
Though we error when the user tries to submit values less than 1, we should also error when the user
submits decimal numbers. This is because we can only send retention time in seconds to the backend.
Furthermore, the backend stores retention time in a long. This means we shouldn't provide decimals in
the request.

**Change**
This diff adds validation for positive integer input before running the minimum value validation for PITR retention
period.

Test Plan:
Test using decimals, negative integers, negative values with decimals, positive integers, positive integers
lower than the expected minimum and positive integers which exceed the minimum.
{F276016}
{F276017}
{F276018}
{F276019}
{F276020}
{F276026}
{F276027}

Reviewers: rmadhavan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37343