[DEVOPS-3048] test automation: Backup unit tests use YBC by default
Summary:
Some unit tests optionally use YB Controller for backup functionality, but assumes you have the
YB Controller SW downloaded. This has been available via environment variable.

This change turns on YBC use by default (for linux), and adds yb_build option to NOT use it.

It also automatically downloads the YBC version specified as the current stable version in the
existing managed/src/main/resources/reference.conf file. The jenkins test script previously used
whatever version was staged on the local machine, but that is also now consolidated to use the same
version as the yb_build.sh uses.

Tarballs are staged in /opt/yb-build/ybc/ so that they are used across build directories without
re-downloading. Once the binaries are unpacked in the local build/ybc directory, they are assumed
to be good. The build/ybc directory can be removed to be sure to pull in new versions.

Test Plan:
 ./yb_build.sh --sb --java-test 'org.yb.pgsql.TestYbBackup#testMixedLegacyColocatedDatabase'
 ./yb_build.sh --sb --no-ybc --java-test 'org.yb.pgsql.TestYbBackup#testMixedLegacyColocatedDatabase'
 YB_TEST_YB_CONTROLLER=0 ./yb_build.sh --sb --java-test 'org.yb.pgsql.TestYbBackup#testMixedLegacyColocatedDatabase'

Also jenkins to test that workflow.

Reviewers: asharma, jharveysmith

Reviewed By: jharveysmith

Subscribers: devops, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37348