[#23286] xCluster: Speedup setup for large table counts
Summary:
Rewrite of `xcluster_universe_replication_setup_helper` to speed up xCluster setup involving large number of tables.
For 1000 tables this change reduces the setup time from 100s to 0.6s.

Most of this improvement is from writing to `sys_catalog` once overrall, instead 2-3 times per table. The old logic stored all intermediate state required by the setup in the `UniverseReplication` COW object. This meant that parallel rpcs were getting serialized by the `UniverseReplication` write to `sys_catalog`. It was also causing a high IO load on the yb-master.

With this change we now store all intermediate state in-memory, and only perform one write to `sys_catalog`. This is safe to do since the xCluster Setup is not resilient to yb-master restarts, so there is no need to persist these states on disk.
The other advantages of this approach is that, on a master crash during setup, we do not leave any stale data that needs to be cleaned up later.
This also helps simplify the ALTER replication, since now there is no more need to store a `.ALTER` `UniverseReplication` object. Everything is kept in memory and we directly make the final updates in the `Original UniverseReplication` object.

The sequences of steps have changed. Most notably we now get the tablet list after setting up the streams and it now uses an async rpc call. Also, we now only get the Partition info which has a smaller subset of `GetTableLocationsResponsePB`.

All Tablegroup related code has been removed, and we now will fail if any table groups are provided. This is because yb does not support tablegroups anymore.

`IsBootstrapRequired` performs a sync call to the source master. This has not been optimized since we do not use this feature. `check_bootstrap_required` flag is defaulted to `false`.

Removed the `WaitForSetupUniverseReplicationToFinish ` in `RemoveDroppedTablesOnConsumer` since the drop is synchronous.

**Upgrade/Rollback safety:**
`schema_version_mappings` has been removed from `SysUniverseReplicationEntryPB`. This was only used in the Setup process, and the process is not retried on master failovers. So it is safe to remove this field.

Fixes #23286
Jira: DB-12211

Test Plan: Jenkins

Reviewers: jhe, xCluster

Reviewed By: jhe

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D36598