[PLAT-14963]Clicking use same/diff replica while DR repair is throwing a permission error for a superadmin user
Summary:
This commit adds the isSuperAdminUser function to RbacUtils. This function checks if the user has super admin privileges by searching for the SUPER_ADMIN_ACTIONS permission in the user's permissions list. If the permission is found, the function returns true; otherwise, it returns false.

Added comments

Test Plan: Tested manually by logging in with super admin and as normal user

Reviewers: lsangappa, jmak

Reviewed By: jmak

Subscribers: jmak, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37356