[PLAT-12905] Add HA Metrics Page
Summary:
There are two gauge metrics that we want to expose to users on the YBA HA page:
- Last backup time
- Last backup size
The last backup time is more useful when presented as a lag metric.
Thus, the UI does an extra processing step to convert the timestamps into lag durations.

This metrics tab makes sense for the active instance when HA replication is set up.
Thus, we hide the metric tab when the current instance is standby or if HA replication is not set up.

Test Plan:
Verify HA metrics are displayed correctly on the admin tab when HA is configured.
{F276928}

Reviewers: amalyshev, rmadhavan

Reviewed By: amalyshev

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35831