[PLAT-14982] Fix restart config table key formatting
Summary:
The row key for tables in the restart config modal was not formatted the same
way as the stored table selection.
The table uuids provided on the backend can have dashes or not.
We strip the dash from all table uuids and use that in xCluster API requests.
As the row key used table uuids with dashs, the UI component did not
recognize the table was in fact selected. This means the UI shows the table
as unselected even though the table is actually selected and gets submitted
to the backend.

The fix in this diff strips the dash from table UUID before handing off the
table to the UI component.

Test Plan:
Verify user restart table selection for YCQL replication configs work as
expected.
{F276457}

Reviewers: rmadhavan, kkannan, lsangappa

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37391