[#23179] CDCSDK: Refactor TestPgReplicationSlot for dynamic data types
Summary:
This revision refactors the test method `replicationConnectionConsumptionAllDataTypes` along with related test framework to improve the interface for creating `PgOutputRelationMessageColumn` for dynamic data types. Since the OIDs of such types are not static, it is better if we do not specify them while creating `PgOutputRelationMessageColumn` for columns with dynamic types.
Jira: DB-12118

Test Plan: Jenkins: test regex: .*TestPgReplicationSlot.*

Reviewers: stiwary, jason

Reviewed By: jason

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37383