[#23543] docdb: Update tablegroup manager in RepartitionTable
Summary:
Previously, the tablegroup manager was not updated when repartitioning a colocated table.
This was causing CREATE TABLE to add tables to the old (deleted) tablet when trying to create tables
on a cloned colocated database. This diff changes `RepartitionTable` to update the tablegroup manager
to point to the new tablet instead of the old one.
Jira: DB-12461

Test Plan: `./yb_build.sh release --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter PgCloneColocationTest.CreateTableAfterClone`

Reviewers: mhaddad, zdrudi

Reviewed By: zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37436