[BACKPORT 2.20][#23394][#22962] CDCSDK: Prevent tserver crash on concurrent Getchanges call on same producer tablet
Summary:
**Backport summary:**
Minor merge conflict as `ProducerTabletInfo` was renamed to `TabletStreamInfo` in latest master.

**Original summary:**
Original commit: 22657daf0a7385ffe2e7da44232d1e938cd4e8f4 / D37269
In CDCService, we maintain a map (`SchemaDetailsMap`) for each tablet, stream pair (referred as `producer_tablet`) in the `cdc_stream_metadata` object. The schema details correspond to the current 'running' schema of the table that was read as of a particular WAL/Intent entry's hybrid_time.

During a Getchanges call on a producer_tablet, we take the reference of its SchemaDetailsMap, which is then modified (if required) and copied back to `cdc_stream_metadata` object. On concurrent GetChanges call on the same `producer_tablet`, we run into a tserver crash since multiple threads are trying to access & modify the same `SchemaDetailsMap` of the producer_tablet. Generally, concurrent GetChanges on same producer_tablet are not expected but they can happen due to retries.

As part of the fix, instead of giving reference, each thread will now be given its own copy of 'SchemaDetailsMap', so that it can be modified independently. Later the cdc_stream_metadata for the producer_tablet will be updated by moving the  modified SchemaDetailsMap copy in the actual map object in a thread-safe manner by taking appropriate locks (existing code - `UpdateCDCStateMetadata()`).

**Proof of correctness**
As these concurrent calls are generated due to retries, all of them are requesting for changes from the same checkpoint. Therefore, even though there are multiple threads modifying the SchemaDetailsMap, we will not lose any updates as all the threads would have read the schema as of the same HT and hence, all the threads are updating the map with the same schema details.
Jira: DB-12316, DB-11880

Test Plan: Jenkins

Reviewers: skumar, stiwary

Reviewed By: stiwary

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37446