Update distributed-transactions-ysql.md
Remove usage of the phrases "optimistic locking" and "pessimistic locking". Because "optimistic locking" doesn't really mean anything and is confused with Optimistic Concurrency Control is a totally different concept in the database community. We were using the phrase "optimistic locking" to point to the semantics of Fail-on-Conflict concurrency control mode and "pessimistic locking" to point to the semantics of Wait-on-Conflict concurrency control.