[PLAT-13757][PLAT-15032][PLAT-15005] Add bootstrap summary page
Summary:
This diff adds a bootstrap summary page to the xCluster modals.
The purpose of this page is to present an overview of which tables are planned for bootstrapping and which are not.
The summary page is also where the user may opt to skip bootstrapping (provided they have enabled this runtime config).
We show a warning to the user if the bootstrapping plan may result in data inconsistencies due to skipped bootstrapping.

Furthermore, we disable the option to restart replication if all tables in the replication config are involve in bidirectional replication.
This is because currently restart replication involves performing a backup & restore which will wipe out data.

Test Plan:
Create xCluster config with tables that require bootstrapping.
- include tables which are missing on target
- include tables which have data
- include tables which are replicated in the opposite direction
- include tables which are empty and present on both sides (no bootstrapping required)
{F278047}
{F278046}
Edit an xCluster config
- likewise test using tables in the 4 categories mentioned above.

Verify that user is not allowed to restart replication if all tables are involved in bidirectional replication.

Reviewers: rmadhavan, hzare, cwang, vbansal

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37105