[BACKPORT 2024.2][#22372] docdb: Backward scans: Speed up block iterator Prev method
Summary:
Block's iterator Prev() call is approximately 5 times slower than a Next() call of the iterator.
This change implements a new approach for Prev() call of the block's iterator to make it faster by
caching of all keys in the current restart block, allowing a random access to any of its entries.

**Background**
SST file consists of blocks, all keys in a block are grouped into sections called 'Restart block'.
All restart blocks are indexed for a given block and can be accessed directly. But entries in a
restart block do not support random access, i.e. it is required to make N reads to reach Nth entry.

Additionally, to reduce a space, required for a restart block, there's a special format for storing
the keys for the entries: the first entry contains a full key, all other entries contain a delta
between the current and the previous keys. Thus, to reach Nth entry in the restart block, it's not
enough to make N reads, but also it's required to apply N-1 deltas to the key of the first entry.

**Backward scan of a restart block**
In case of a backward scan over the whole restart block, it is required to read all its entries
in the reverse order. Taking into account how the data is structured and persisted in a restart
block, the following approach is used:
1.. To read N-th entry of the restart block, it's required to read all entries from the very first
entry upto Nth entry and to apply all N-1 key deltas to the first key.
2.. To read N-1-th entry of the restart block, it's required to read all entries from the very first
entry upto N-1-th entry and to apply all N-2 key deltas to the first key.
and so on ...

It is clearly seen that there is a lot of repeated work (read and apply deltas) done to get
the I-th key of a restart block.

**New approach**
The main idea of the new approach is to avoid that repeated work and keep (cache) all the full keys
for the current restart block during the first forward pass to the Nth key. Such an approach allows
a random access to any key of the current restart block which saves CPU cost on the reading and
applying key deltas.
The **trade off**: additional memory is required to copy and store 16 keys instead of 1 key. This
memory is released once the iterator moves to the next restart block.

The approach is enabled by default, if fast backward fast is enabled (use_fast_backward_scan).
It can be disabled by setting rocksdb_iterator_disable_restart_block_keys_caching to true.

**Perf result**
RocksDB benchmark results run on a standard dev server in release mode,
num of keys: 100000, num of iterations: 1000
|                   |      baseline      | revision, no cache | revision, with cache |
| forward scan, us  |             4545.8 |             4463.9 |                    - |
| backward scan, us |            25946.2 |            25989.4 |               5163.6 |
Jira: DB-11273

Original commit: 18685016a2f280bcc3a46c07e9e98c2020e887f7 / D37464

Test Plan:
Jenkins

To test data correctness during iteration:
./yb_build.sh  --cxx-test='TEST_F(HarnessTest, Randomized)'
./yb_build.sh  --cxx-test='TEST_F(HarnessTest, RandomizedLongDB)'

To run RocksDB benchmarks:
1.. Enable `BlockBasedTableTest.FastBackwardScanPerformanceTest` by removing `DISABLED_` from
its declaration `TEST_F(BlockBasedTableTest, DISABLED_FastBackwardScanPerformanceTest)`
2.. Run `./yb_build.sh --cxx-test='TEST_F(BlockBasedTableTest, FastBackwardScanPerformanceTest)'`
3.. Repeat 1-2 for other required benchmarks such as BackwardScanPerformanceTest,
ForwardScanPerformanceTest, ScanPerformanceTest, etc.

Reviewers: timur, rthallam

Reviewed By: rthallam

Subscribers: ybase, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D38646