[#23613] DocDB: Framework for different vector index coordinate types, SIFT 1B hnsw_tool support
Summary:
A framework for supporting different vector coordinate types. Many vector-related classes have to become templates parameterized with the coordinate type. In this diff, they are actually parameterized with the vector type, not the coordinate type, where the vector type must satisfy the IndexableVectorType concept, because we need to work with vectors, not with coordinates, most of the time.

Usearch currently does not support u8 (unsigned byte) data type directly, so we are still converting vectors to float before invoking usearch. We are still able to run some tests with the SIFT 1B dataset with this setup.

We are also doing significant refactoring in hnsw_tool as part of making sure we can run tests on the SIFT 1B dataset.
Jira: DB-12525

Test Plan:
Jenkins

Manual testing:

  bin/hnsw_tool benchmark \
    --build_vecs_path sift1b/bigann_base.bvecs \
    --query_vecs_path sift1b/bigann_query.bvecs \
    --ground_truth_path sift1b/gnd/idx_5M.ivecs \
    --num_vectors_to_insert=5000000 \
    --k=1000 \
    --num_indexing_threads=15 \
    --num_query_threads=80

Reviewers: sergei, aleksandr.ponomarenko

Reviewed By: sergei

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37459