[BACKPORT 2.20][#23518] docdb: Do not include full snapshot info for list snapshot schedules RPC.
Summary:
The SnapshotInfoPB included in the ListSnapshots and ListSnapshotSchedules RPC responses includes a copy of the SysSnapshotEntryPB proto stored in the sys catalog. This proto has an entry for every tablet and table in the database snapshotted. This can be quite large for larger dbs. Most clients are not interested in this detail. To limit the response size we've already done [[ https://github.com/yugabyte/yugabyte-db/issues/13633 | work ]] on ListSnapshots to omit the per-object details from the included SysSnapshotEntryPB. However we neglected to modify ListSnapshotSchedules when we did this work. We've seen instances where the response to ListSnapshotSchedules is so large it overflows the default response size of 245MiB, breaking the RPC.

This diff omits the entries information of SysSnapshotEntryPB from ListSnapshotSchedules. This is a breaking change but we're not aware of any callers of ListSnapshotSchedules that use this information. Furthermore this information is still obtainable from `ListSnapshots` with the right parameters.

In terms of the proto schemas:
Before this diff, this field in the `ListSnapshotSchedulesResponse` was populated:
```
ListSnapshotSchedulesResponsePB::schedules::snapshots::entry::entries
```
After this diff it is not populated.

This is a breaking change but I audited the YBA code that invokes snapshot schedule APIs and none of it accesses the `entry` field in the returned `SnapshotInfoPB`s. We may break open source user workflows but given how easy it is to misuse `ListSnapshotSchedules` this change is worth the cost.

Original commit: 1e28b8a07f0519393a7b241c9a619699a6029bba / D37359
Jira: DB-12436

Test Plan:
```
% ./yb_build.sh --with-tests --cxx-test minicluster-snapshot-test --gtest_filter 'MasterSnapshotTest.ListSnapshotSchedulesHasNoSnapshotDetails'
```

Reviewers: asrivastava, mhaddad, vpatibandla

Reviewed By: asrivastava

Subscribers: ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D37547