[PLAT-15061] Update integration page naming
Summary:
This is a simple diff that updates the provider configurations page title to `Integrations` to be in line
with the new tab page.

The diff also adds the i18n for the string on this page.

Test Plan:
Verify the YBA integrations page loads fine and all texts appear as expected.
{F278040}

Reviewers: rmadhavan, agarg

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37515