[PLAT-14988] Preselect index tables if the main table is selected
Summary:
**Context**
YBA backend has logic to add index tables in the main table is selected and `autoIncludeIndexTables` is true.
But it is clearer for the user if we simply add the index tables to the table selection from the client side.
Otherwise, customers may be confused the index table doesn't appear as part of the selected table count even though
it gets added by the backend upon form submission.

**Changes**
- Allow selecting index tables when the user clicks the select unreplicated tables button
- Preselect index tables if the main table is selected

Test Plan:
- Create a transactional xCluster config
- Add index tables on the source but not the target.
- Verify that these tables get preselected in the 'edit tables modal'.
  - If the modal is opened immediately after the index table is added, then
    the newly added table might not be preselected. In that case, clicking the
    add unreplicated tables button will add then index table.
- Verify the table count in both edit table modal and restart config modal reflect the selected index tables as well.
{F277599}

Reviewers: rmadhavan, hzare, cwang, vbansal

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37491